Tile as u32 (and configurable)!

+++ src/base.rs
@@ -6,7 +6,7 @@ use thiserror::Error;
 pub type Point = (usize, usize);
 pub type NumTiles = u32;
 pub type NumEvents = u64;
-pub type Tile = usize;
+pub type Tile = u32;
 pub type Energy = f64;
 pub type Glue = usize;
 pub type CanvasLength = usize;
